perf: skip unifying equal labels
On a small test on 512x512x512 connectomics data at 32x32x40
resolution, this brought down the time from 12 seconds to 7.
SciPy (which regarded the volume as binary) took around 5.9,
so we are moving from the 1980s or 1990s into the 2000s in
terms of performance. Using Wu or Grana would be even more
helpful.