feat: support arbitrary labels
Using unordered map instead of an array is several times slower,
but we can support the use of IDs that are larger than the array
size. Right now, one must run renumber on arbitrary data before
running cc3d. The array also takes up a lot of space.

For ordinary data, the hashmap is more memory efficient. However,
on the pathological case of sequence data, memory blows up.

Ideally, you should be able to choose which version of union-find
to use.