perf: recover performance on workstation
Converting the loc->x,y,z logic into x,y,z->loc was faster on my
laptop using clang, but slower on my workstation using g++. gprof
showed the for loop was now the bottleneck. I didn't disassemble
the code, but I suspected that either the code generated was not
great or the ALU pipeline was saturated. Previously, there had been
some load on the divisor unit which might have given a break to the
adder/multiplier and allowed more throughput there.

I tried reducing the width of the x,y,z variables and this seemed to
help on the 26 connected code, so something was getting saturated at
least.

I doubt that anyone will be creating 1D images that need a 64 bit
integer (or even a 32 bit integer), but I would be happy to revisit
this decision if someone does. Most people make e.g. 2k x 2k x 1k
images, which only need a 16-bit integer in x,y,z but a 64-bit integer
in index.