perf: use phantom label technique to speed up 8-connected
In the typical Wu et al SAUF mask, all pixels included in the mask
are eligible to be unified with or to have their value stolen.

Here we introduce what I term "phantom labels" which are labels
that stand outside the standard mask, are not eligible for a
unify operation, but still transmit connectivity information.

Here we use the existence of phantom label P to
sometimes skip a unify operation. In this case, P allows us
to bridge the gap between A and C when B is missing.

  P
A B C
D E